Handle reflect.Values that are not addressable
map[string]Message does not allow reflect.Value#Addr() on the
reflect.Value retrieved by accessing the map at an index, because the
map value is not accessable as a pointer. Create a copy of the value if
CanAddr() is false before invoking proto.Text and jsonpb.Marshal.

Required to support nullable=false map values