Allow jsonpb to handle cast map values that are not proto.Message
jsonpb doesn't have access to the original type of the cast map value
from Proto, which means the generic struct reflection marshalling fails
if the cast type is not proto.Message. Add a struct tag annotation that
indicates the original message type so we can look it up in the proto
registry.