gogoproto: gostyle enum and custom enum name extensions
This provides a minimal pass at adding support for custom enum names or just
idiomatic, Go-style const names in generated output.

If `enum_gostyle` for an enum type or `enum_gostyle_all` for a file are
enabled, the enum names will have the underscores replaced with a camelcase
variation. If that is insufficient, a custom name can be generated using the
`enum_customname` extension on `EnumValueOptions`, which works identical to the
field targeted variation.

Please evaluate the approach and I'll add unit tests and documentation if this
is acceptable.

Signed-off-by: Stephen J Day <stephen.day@docker.com>