gogoproto: enum and enum value custom name extension
This provides support for custom enum and enum value names allowing one
to override default naming to their delight. The `enum_customname`
extension will allow one to set the type name of an enum. If
`goproto_enum_prefix` is enabled, this value will be used as a prefix
when generating enum values. If `enumvalue_customname` is enabled for a
particular value, the prefix is ignored and the name is used directly.

A new extension set is declared for `EnumValueOptions` in the 66xxx
range to continue with the extension numbering convention.

Signed-off-by: Stephen J Day <stephen.day@docker.com>