Support unsafe borrow unmarshaler
When Go converts a []byte to a string it creates a copy to guarantee
immutability of the string. Since many objects contain strings it is not
uncommon for up to half of all allocations during unmarshal to be the
copy of []byte data to a string.

Go currently implements a non-moving garbage collector - once in place,
pointers are not moved. This commit adds a "borrow" unmarshaler that
assumes the Unmarshal() caller guarantees the immutability of the
incoming data (it will not be modified after the call to Unmarshal) and,
instead of converting []bytes to String during unmarshal, uses an
unsafe.Pointer and reflect.SliceHeader/StringHeader to modify a string
to point directly to the bytes in the incoming data. By doing so the
caller trades many small string allocations for an extra copy up front
of the larger buffer.

For example, an object with 3 nested arrays that has 25 strings and
those strings are 50% of the size of the incoming data will trade:

    Size(strings) = 400 bytes
    Size(serialized object) = 800 bytes
    Size(object) = Size(structs) + 400 bytes
    Allocations(object) = 25 + Allocations(structs)

for

    Size(object) = Size(structs) + 800 bytes
    Allocations(object) = 1 + Allocations(structs)

Generally, halving the allocations during unmarshal results in a 25%
speedup.

If the guarantees of the collector in Go change, this optimization may
no longer function.