Drop types declaration option (#250)
* [FEATURE] Add options for dropping type declarations

Add two options, one for messages and one for enums, to drop type
declarations. In messages, it also drops getters and setters. In enums,
it also drops the generated constants.

* [GENERATED] Add options for dropping the type

* Fix drop type decl tests Makefile

* Fix protoc calls

* Rename test proto files and fix makefiles.

* Make drop type declaration and go getters options independent

This commit extract a few variable initializations that was being done
at the same time of generating the type for a given message. This allow
for the Go Getters option to be independent from the drop type declaration
one

* Readd comment

* [FEATURE] Rename options and add file-scoped versions

* `drop_type_declaration` becomes `typedecl`
* `enum_drop_type_declaration` becomes `enumdecl`

Default values are now true and for removing the declaration the value needs to
be `false`.

Options `typedecl_all` and `enumdecl_all` are added.

* [GENERATED] Side effects of previous commit

* Improve documentation

* Add me to contributors