Add references to some custom imports
When using stdtime and typedecl options, an error is raised because we are
importing time and not using it (if the time field is not nullable).

This is fixed substituting `customImports` in the generator to receive a custom
struct containing three optional fields:

* *Alias* for the import
* *Path* for the import
* *UnderscorableSymbol* for an exported symbol in the package that can be used
as an expression (it will be used in `var _ = <Alias>.<UnderscorableSymbol>`)

Then, `writtenImports` is used to make sure we don't write the `var` assignment
twice.

If a `customImport` has no `UnderscorableSymbol`, then the statement won't be
generated at all.