plugin/populate: avoid loops on non top-level packages
The previous scheme for detecting when a type belonged to a different
package only considered the topmost component of the type name, but this
is not a sufficiently general means of determining a type's package, as
package names can themselves contain dots. Get the package name from the
file descriptor instead, which is guaranteed to be correct.

Also add a test to ensure this doesn't break in the future.

Fix #207 (again).