Adding new plugin protoc-gen-gogo-gateway
This plugin fixes an issue where gogoproto.customname is not compatible with
protoc-gen-grpc-gateway, when this field is also referenced as a path parameter.
This happens frequently with id field of an object. Generator by default uses
CamelCase (Id) while go advises to ID to be used. gogoproto.customname helps
the generator to adjust the fieldname, but protoc-gen-grpc-gateway does not know
the field is altered and keep using CamelCase version (Id).